Replace kubectl wait with wait_for function
The `kubctl wait --for=... --timeout=...` command is almost useful, but it
suffers from a fatal flaw: if the requested path doesn't exist *yet*, it
will fail immediately rather than waiting for the path to become available.
This means that in almost all cases we can't use the `--timeout` option.

This commit replaces the `kubectl wait --timeout...` command with the
`wait_for` function, which wraps `kubectl wait` (without `--timeout`) in a
while loop and uses the `timeout` command to implement the timeout.